Moved application code to new directory economicalc
