+++ backend/economicalc/objects/image.py
